Potential fix for code scanning alert no. 21

+++ core/drivers/firewall/stm32_etzpc.c
@@ -209,7 +209,8 @@ etzpc_do_get_decprot(struct etzpc_device *etzpc_dev, uint32_t decprot_id)
 static void etzpc_do_lock_decprot(struct etzpc_device *etzpc_dev,
 				  uint32_t decprot_id)
 {
-	size_t offset = U(4) * (decprot_id / IDS_PER_DECPROT_LOCK_REGS);
+	size_t offset = (size_t)U(4) *
+			(decprot_id / IDS_PER_DECPROT_LOCK_REGS);
 	uint32_t mask = BIT(decprot_id % IDS_PER_DECPROT_LOCK_REGS);
 	vaddr_t base = etzpc_dev->pdata.base.va;
 	uint32_t exceptions = 0;
